Added files: /userprog/car.cpp, /userprog/car.h, /userprog/car_proto.cpp, /userprog/car_proto.h NOTE: They lack documentation but it will be added in a later commit.
They correspond to (car.h/.cpp) a class that serves a thread for the bridge problem. It is
intended to use a function pointer to the OneVehicle() function (see car_proto) to execute.

car_proto defines the functions OneVehicle(), ArriveBridge(), CrossBridge(), ExitBridge()
as well as references to several external variables that will be necessary for the
synchronization of the car (see car_proto.h), the names should be self explanatory.